#6: Creating registration.types module to hold general-use data structures
